feature gate some lossful conversion
