+++ crates/scroll/revm/src/states/reverts.rs
@@ -110,18 +110,6 @@ impl From<(AccountRevert, &ScrollPostExecutionContext)> for ScrollAccountRevert 
     }
 }
 
-// This conversion can cause a loss of information since performed without additional context.
-impl From<AccountRevert> for ScrollAccountRevert {
-    fn from(account: AccountRevert) -> Self {
-        Self {
-            account: account.account.into(),
-            storage: account.storage,
-            previous_status: account.previous_status,
-            wipe_storage: account.wipe_storage,
-        }
-    }
-}
-
 impl ScrollAccountRevert {
     /// The approximate size of changes needed to store this account revert.
     /// `1 + storage_reverts_len`
@@ -160,17 +148,6 @@ impl From<(AccountInfoRevert, &ScrollPostExecutionContext)> for ScrollAccountInf
             AccountInfoRevert::DoNothing => Self::DoNothing,
             AccountInfoRevert::DeleteIt => Self::DeleteIt,
             AccountInfoRevert::RevertTo(account) => Self::RevertTo((account, context).into()),
-        }
-    }
-}
-
-// This conversion can cause a loss of information since performed without additional context.
-impl From<AccountInfoRevert> for ScrollAccountInfoRevert {
-    fn from(account: AccountInfoRevert) -> Self {
-        match account {
-            AccountInfoRevert::DoNothing => Self::DoNothing,
-            AccountInfoRevert::DeleteIt => Self::DeleteIt,
-            AccountInfoRevert::RevertTo(account) => Self::RevertTo(account.into()),
         }
     }
 }
